% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dm.R
\name{fit_dm}
\alias{fit_dm}
\title{Fit a diachronic model}
\usage{
fit_dm(
  data,
  model,
  reps = 10,
  sdfactor = 0.01,
  method = "BFGS",
  control = list(),
  mc.cores = round(parallel::detectCores()/2)
)
}
\arguments{
\item{data}{Data set in long format, with columns \code{time}, \code{context} and \code{response}}

\item{model}{Model to fit: see Details for options}

\item{reps}{Number of repetitions to take (with different initial parameter guesses)}

\item{sdfactor}{Initial parameter guess S.D. prefactor}

\item{method}{Optimization method to use}

\item{control}{Control parameters to optimization routine}

\item{mc.cores}{Number of parallel processor cores to use (Mac and Linux only)}
}
\value{
Data frame with following components:
\describe{
\item{model:}{Model fitted}
\item{N:}{Number of data points}
\item{C:}{Number of contexts}
\item{K:}{Number of model parameters}
\item{method:}{Numerical optimization method employed}
\item{reps:}{Number of fitting repetitions}
\item{runtime:}{Runtime (wall-clock) of fitting}
\item{convergence:}{Convergence information returned by \code{optim}: \code{0} indicates convergence}
\item{logL:}{Maximum of log-likelihood function found}
}
}
\description{
Fit a constant or variable rate effect model to data using a maximum likelihood method.
}
\details{
Currently available models are \code{"CRH"} and \code{"VRH"} for the classical Constant and Variable Rate Hypotheses (Kroch 1989), \code{"qCRH"} and \code{"qVRH"} for their quadratic extensions (Kallel 2007), and \code{"BRH"} for the Biased Rate Hypothesis (Kauhanen and Walkden 2018). See Kauhanen (in prep) for details of these, including calculations for the likelihood functions and their gradients.

The maximum likelihood estimate is found using \code{optim}. By default, the BFGS algorithm is assumed (this seems to perform the best across the board), though other methods may be specified through the \code{method} (see the \code{optim} documentation for the available options). The behaviour of the optimization algorithm can be further controlled through use of the \code{control} parameter.

An initial parameter guess is found by fitting a logistic model of the form \code{response~time} without regard for contexts. To avoid convergence to local optima, the optimization is repeated from \code{reps} such initial guesses, drawing each parameter from a normal distribution whose mean is the parameter found by the contextless logistic regression and whose standard deviation equals \code{sdfactor} times that mean; coefficient estimates for the quadratic terms in the qCRH and qVRH models, as well as those for the bias parameters in the BRH model, however, start with zero estimates as this seems to best guarantee convergence. The repetitions are conducted in parallel using \code{mclapply} if a multicore processor is present and forking is possible (i.e. on Linux and Mac systems). Increasing the value of \code{reps} will lessen the probability of idiosyncratic fits but will increase computation time.
}
\references{
Kallel, Amel (2007) The loss of negative concord in Standard English: internal factors. \emph{Language Variation and Change}, 19, 27--49.

Kauhanen, Henri (in prep) Grammar competition, speaker models and rates of change: a critical reappraisal of the Constant Rate Hypothesis.

Kauhanen, Henri & Walkden, George (2018) Deriving the Constant Rate Effect. \emph{Natural Language & Linguistic Theory}, 36, 483--521.

Kroch, Anthony S. (1989) Reflexes of grammar in patterns of language change. \emph{Language Variation and Change}, 1, 199--244.
}
\seealso{
\code{optim}
}
