% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dm_all.R
\name{fit_dm_all}
\alias{fit_dm_all}
\title{Fit All Available Diachronic Models}
\usage{
fit_dm_all(
  data,
  models = list("CRH", "VRH", "qCRH", "qVRH", "BRH"),
  reps = 10,
  sdfactor = 0.01,
  method = "BFGS",
  control = list(),
  mc.cores = round(parallel::detectCores()/2)
)
}
\arguments{
\item{data}{Dataset in long format}

\item{models}{Models to fit; by default, all available models are fit}

\item{reps}{Number of repetitions to take (with different initial parameter guesses)}

\item{sdfactor}{Initial parameter guess S.D. prefactor}

\item{method}{Optimization method to use}

\item{control}{Control parameters to optimization routine}

\item{mc.cores}{Number of parallel processor cores to use (Mac and Linux)}
}
\value{
List with the following elements:
\describe{
\item{summary:}{Summary of fit; a dataframe combining the individual outputs for each model}
\item{coef:}{Maximum likelihood estimates of coefficients}
}
}
\description{
Fit all currently available models to a dataset.
}
\details{
For further explanation of the fitting routine, see \code{fit_dm}.
}
